requirements.txt
# Fabric GSM Prediction Pipeline Requirements
# Python 3.10+

# Core ML and Image Processing
tensorflow>=2.13.0
keras>=2.13.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
scikit-image>=0.21.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0

# Configuration Management
pyyaml>=6.0

# Utilities
pillow>=10.0.0